Add utility hasPropertyWithName
