+++ src/transforms/v2-to-v3/utils/add/addV3ClientRequires.ts
@@ -7,20 +7,9 @@ import {
   getV3ClientRequireProperty,
   getV3ClientTypeNames,
 } from "../get";
+import { hasPropertyWithName } from "../has";
 import { addV3ClientModuleRequire } from "./addV3ClientModuleRequire";
 import { AddV3ClientModulesOptions } from "./addV3ClientModules";
-
-const hasIdentifierName = (varDeclarator: VariableDeclarator, localName: string) =>
-  varDeclarator.id.type === "Identifier" && varDeclarator.id.name === localName;
-
-const hasObjectPropertyName = (varDeclarator: VariableDeclarator, localName: string) =>
-  varDeclarator.id.type === "ObjectPattern" &&
-  (varDeclarator.id as ObjectPattern).properties.some(
-    (property) =>
-      property.type === "Property" &&
-      property.value.type === "Identifier" &&
-      property.value.name === localName
-  );
 
 export const addV3ClientRequires = (
   j: JSCodeshift,
@@ -45,26 +34,11 @@ export const addV3ClientRequires = (
     // Insert after require for global SDK if present. If not, insert after service require.
     const v2ServiceModulePath = getV2ServiceModulePath(v2ClientName);
     const globalRequireVarDecl = getRequireVariableDeclaration(j, source, PACKAGE_NAME).filter(
-      (varDeclaration) => {
-        const declarations = varDeclaration.value.declarations.filter(
-          (declaration) => declaration.type === "VariableDeclarator"
-        ) as VariableDeclarator[];
-
-        if (
-          v2GlobalName &&
-          declarations.some((declaration) => hasIdentifierName(declaration, v2GlobalName))
-        ) {
-          return true;
-        }
-
-        if (
-          declarations.some((declaration) => hasObjectPropertyName(declaration, v2ClientLocalName))
-        ) {
-          return true;
-        }
-
-        return false;
-      }
+      (varDeclaration) =>
+        hasPropertyWithName(varDeclaration, {
+          identifierName: v2GlobalName,
+          objectPropertyName: v2ClientLocalName,
+        })
     );
     const serviceRequireVarDecl = getRequireVariableDeclaration(j, source, v2ServiceModulePath);
