+++ src/transforms/v2-to-v3/utils/has/index.ts
@@ -1,1 +1,2 @@
+export * from "./hasPropertyWithName";
 export * from "./hasRequire";
