+++ src/transforms/v2-to-v3/utils/remove/removeRequireIdentifierName.ts
@@ -1,23 +1,12 @@
-import { Collection, JSCodeshift, ObjectPattern, VariableDeclarator } from "jscodeshift";
+import { Collection, JSCodeshift } from "jscodeshift";
 
 import { getRequireVariableDeclaration } from "../get";
+import { hasPropertyWithName } from "../has";
 
 export interface RemoveRequireIdentifierNameOptions {
   localName: string;
   sourceValue: string;
 }
-
-const hasIdentifierName = (varDeclarator: VariableDeclarator, localName: string) =>
-  varDeclarator.id.type === "Identifier" && varDeclarator.id.name === localName;
-
-const hasObjectPropertyName = (varDeclarator: VariableDeclarator, localName: string) =>
-  varDeclarator.id.type === "ObjectPattern" &&
-  (varDeclarator.id as ObjectPattern).properties.some(
-    (property) =>
-      property.type === "Property" &&
-      property.value.type === "Identifier" &&
-      property.value.name === localName
-  );
 
 export const removeRequireIdentifierName = (
   j: JSCodeshift,
@@ -25,15 +14,11 @@ export const removeRequireIdentifierName = (
   { localName, sourceValue }: RemoveRequireIdentifierNameOptions
 ) => {
   getRequireVariableDeclaration(j, source, sourceValue)
-    .filter((variableDeclaration) => {
-      const declarations = variableDeclaration.value.declarations as VariableDeclarator[];
-      if (declarations.some((varDeclarator) => hasIdentifierName(varDeclarator, localName))) {
-        return true;
-      }
-      if (declarations.some((varDeclarator) => hasObjectPropertyName(varDeclarator, localName))) {
-        return true;
-      }
-      return false;
-    })
+    .filter((varDeclaration) =>
+      hasPropertyWithName(varDeclaration, {
+        identifierName: localName,
+        objectPropertyName: localName,
+      })
+    )
     .remove();
 };
